feat: add page life cycle
