+++ packages/components/src/layouts/Page/PageLifeCycle.tsx
@@ -1,46 +1,8 @@
-import { useEffect } from 'react';
+import { usePageLifeCycle } from './hooks';
 
-import { useNavigation } from '@react-navigation/core';
-
-export interface IPageLifeCycle {
-  onMounted?: () => void;
-  onUnmounted?: () => void;
-}
-
-export function usePageLifeCycle(params?: IPageLifeCycle) {
-  const navigation = useNavigation();
-  const { onMounted, onUnmounted } = params || {};
-  useEffect(() => {
-    if (!onMounted || !onUnmounted) {
-      return;
-    }
-    const unsubscribe = navigation.addListener('transitionEnd' as any, (e) => {
-      const {
-        data: { closing },
-      } = e as {
-        data: {
-          closing: boolean;
-        };
-        target: string;
-        type: string;
-      };
-
-      if (closing) {
-        onUnmounted?.();
-      } else {
-        onMounted?.();
-      }
-    });
-
-    return () => {
-      setTimeout(() => {
-        unsubscribe();
-      }, 100);
-    };
-  }, [navigation, onMounted, onUnmounted]);
-}
+import type { IPageLifeCycle } from './type';
 
 export function PageLifeCycle({ onMounted, onUnmounted }: IPageLifeCycle) {
-  usePageLifeCycle();
+  usePageLifeCycle({ onMounted, onUnmounted });
   return null;
 }
