+++ packages/components/src/layouts/Page/index.tsx
@@ -13,14 +13,14 @@ import {
   FooterConfirmButton,
 } from './PageFooterActions';
 import { PageHeader } from './PageHeader';
-import { PageLifeCycle, usePageLifeCycle } from './PageLifeCycle';
+import { PageLifeCycle } from './PageLifeCycle';
 
 import type { IPageFooterRef } from './PageContext';
 import type { IPageProps } from './type';
 import type { IScrollViewRef } from '../ScrollView';
 import type { NativeScrollPoint } from 'react-native';
 
-export type { IPageProps, IPageFooterProps } from './type';
+export type { IPageProps, IPageFooterProps, IPageLifeCycle } from './type';
 
 function PageProvider({
   children,
